requirements.txt
# Interface web
streamlit>=1.32.0

# Traitement d'images
opencv-python>=4.5.0
Pillow>=9.2.0

# Moteurs OCR
pytesseract>=0.3.10
easyocr>=1.6.0
python-doctr[torch]>=0.12.0

# Traitement de documents
python-docx>=0.8.11
PyMuPDF>=1.18.19
openpyxl>=3.0.0

# Machine Learning
torch>=2.0.0
torchvision>=0.15.0

# Utilitaires
pandas>=2.0.0
numpy>=1.21.0
scipy>=1.4.0
psutil>=5.8.0